Add New Methods and Rename Attribute
To support importing data from JSON, the Day class now contains a
class method load() that accepts a dictionary and converts it to
a Day object if all of the keys in the dict match the attributes
of the Day class. If the imported data is invalid, the class
returns None.

Also added modifyEntry instance method to support updating the
journal entry via a method call. This allows classes that use
Day objects to modify the journal entry via a method call
rather than accessing the Entry instance attributes. The
purpose is to keep the class as modular as possible.

Lastly, the journal_entry attribute was renamed to log.

test_day.py contains tests for each of the specified changes
above.