Modify Instance Methods
Instance method modifyEntry is now modifyLog. This better aligns
with the instance attribute log.

New instance method viewLog was added for access to log content
outside of Day object.

Update to __init__: Log is set to a  generic Entry() if
Entry.load() fails and returns None.

test_day.py was updated to support the above changes.